fastapi/requirements.txt
fastapi[standard]
uvicorn
databases[asyncpg]
pydantic
aiofiles
pyjwt
bcrypt
python-multipart